% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeKi.R
\name{computeKi}
\alias{computeKi}
\title{function to compute Ki, when EC50 is already available, and insertedHotRadioligandConcentration and Kd are provided.}
\usage{
computeKi(EC50, insertedHotRadioligandConcentration, Kd)
}
\arguments{
\item{EC50}{EC50 to be transformed to Ki}

\item{insertedHotRadioligandConcentration}{inserted hot radioligand concentration}

\item{Kd}{Kd}
}
\value{
Ki
}
\description{
function to compute Ki, when EC50 is already available, and insertedHotRadioligandConcentration and Kd are provided.
}
\author{
Vahid Nassiri
}
