% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empiricialED.R
\name{empiricialED}
\alias{empiricialED}
\title{function to find the empirical effective dose}
\usage{
empiricialED(data, respLevel)
}
\arguments{
\item{data}{a dataset with dose and response values}

\item{respLevel}{the level of response for which we want the concentration}
}
\value{
empirical ED
}
\description{
function to find the empirical effective dose
}
\author{
Vahid Nassiri
}
