% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitLogisticModel.R
\name{fitLogisticModel}
\alias{fitLogisticModel}
\title{function to fit logistic dose-response model based on Dotmatics implementation}
\usage{
fitLogisticModel(
  fitData,
  fixedMin = NA,
  fixedMax = NA,
  fixedSlope = NA,
  useWeights = FALSE,
  insertedHotRadioligandConcentration = NA,
  Kd = NA,
  maxiter = 50,
  ftol = sqrt(.Machine$double.eps),
  xlabel = "Concentration (uM)",
  ylabel = "\%Activation"
)
}
\arguments{
\item{fitData}{a data frame with two columns, first one is the dose and second one is the reponse.}

\item{fixedMin}{default is NA, but if the min of the model is fixed, the fixed value should be provided.}

\item{fixedMax}{default is NA, but if the max of the model is fixed, the fixed value should be provided.}

\item{fixedSlope}{default is NA, but if the slope of the model is fixed, the fixed value should be provided.}

\item{useWeights}{logical parameter specifying whether Tukey's biweight method should be used to give outliers less weight (TRUE) or not (FALSE). Default value is FALSE.}

\item{insertedHotRadioligandConcentration}{it the inserted hot radioligand concentration which is used to compute Ki. When this argument and Kd are given, Ki will also be computed.}

\item{Kd}{Kd which is needed for Ki computation. If this argument and insertedHotRadioligandConcentration are provided then Ki is computed.}

\item{maxiter}{maximum number of iterations in Levenberg-Marquardt Nonlinear Least-Squares Algorithm}

\item{ftol}{non-negative numeric. ftol measures the relative error desired in the sum of squares.}

\item{xlabel}{label of the x-axis in the plot}

\item{ylabel}{label of the y-axis in the plot}
}
\value{
list of components
}
\description{
function to fit logistic dose-response model based on Dotmatics implementation
}
\author{
Vahid Nassiri
}
